refactor: Skeleton 컴포넌트 공통 폴더 이동 (#197)
