+++ app/(main)/log/[id]/page.tsx
@@ -14,6 +14,7 @@ import ControlLayout from '@/components/common/ControlLayout'
 import ErrorMessage from '@/components/common/ErrorMessage'
 import Modal from '@/components/common/Modal'
 import { ModalMessageType } from '@/components/common/Modal/types'
+import DetailSkeleton from '@/components/common/Skeleton/DetailSkeleton'
 import { API_ENDPOINTS } from '@/constants/api'
 import { vehicleService } from '@/lib/apis'
 import { addSpaceVehicleNumber } from '@/lib/utils/string'
@@ -22,7 +23,6 @@ import { breakPoints } from '@/styles/theme.css'
 import '@mantine/dates/styles.css'
 import 'dayjs/locale/ko'
 
-import DetailSkeleton from './components/DetailSkeleton'
 import { useDetailData } from './hooks/useDetailData'
 import { useDetailModal } from './hooks/useDetailModal'
 import { useSearchDate } from './hooks/useSearchDate'
