+++ app/(main)/log/page.tsx
@@ -12,6 +12,7 @@ import ErrorMessage from '@/components/common/ErrorMessage'
 import SearchInput from '@/components/common/Input/SearchInput'
 import Modal from '@/components/common/Modal'
 import { ModalMessageType } from '@/components/common/Modal/types'
+import LogSkeleton from '@/components/common/Skeleton/LogSkeleton'
 import ListHeader from '@/components/domain/vehicle/ListHeader'
 import { LOG_TITLES } from '@/constants/listHeader'
 import { useModal } from '@/hooks/useModal'
@@ -19,7 +20,6 @@ import { validateSearchTerm } from '@/lib/utils/validation'
 import '@mantine/notifications/styles.css'
 
 import LogListItem from './components/LogListItem/index'
-import LogSkeleton from './components/LogSkeleton'
 import { useLogData } from './hooks/useLogData'
 import * as styles from './styles.css'
 import { downloadExcel } from './utils/excel'
