+++ components/common/Skeleton/DetailSkeleton/index.tsx
