+++ components/common/Skeleton/InspectionSkeleton/index.tsx
