+++ components/common/Skeleton/InspectionSkeleton/styles.css.ts
