+++ components/common/Skeleton/NoticeSkeleton/index.tsx
