+++ components/common/Skeleton/NoticeSkeleton/styles.css.ts
